dom/bindings/test/test_exceptionThrowing.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=847119
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 847119</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 847119 **/

  var xhr = new XMLHttpRequest();
  var domthrows = function() { xhr.open();  }

  var count = 20000;
  
  function f() {
    var k = 0;
    for (var j = 0; j < count; ++j) {
      try { domthrows(); } catch(e) { ++k; }
    }
    return k;
  }
  function g() { return count; }

  is(f(), count, "Should get count exceptions");
  for (var h of [f, g]) {
    try { is(h(), count, "Should get count exceptions here too"); } catch (e) {}
  }
  ok(true, "We should get here");

  var domthrows = function() { xhr.withCredentials = false; }
  xhr.open("GET", "");
  xhr.send();

  is(f(), count, "Should get count exceptions from getter");
  for (var h of [f, g]) {
    try { is(h(), count, "Should get count exceptions from getter here too"); } catch (e) {}
  }
  ok(true, "We should get here too");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=847119">Mozilla Bug 847119</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
